requirements.txt
einops==0.6.0
graphviz==0.16
imgaug==0.4.0
matplotlib==3.2.2
numpy==1.21.5
opencv_python==4.4.0.44
Pillow==9.4.0
PyYAML==6.0
scipy==1.7.3
terminaltables==3.1.10
timm==0.6.5
torch==1.8.1+cu111
torchinfo==1.7.2
torchsummary==1.5.1
torchvision==0.9.1
torchviz==0.0.2
tqdm==4.47.0
yacs==0.1.8
